This is a very large hopefully non-breaking refactor of the calendar helper.
The main reason for this is to hopefully improve readability and
separate out functionality into more meaningful methods, reduce the
comment count and increase the prospect of getting some retroactive
unit tests on this.

The main approach has been the extract method refactor. Wherever there
was a comment in the code that would be better explained by a named
method I chose to do that.
Moving into lots of methods meant that we still are tied to holding
state in what were local variables but are now method calls to small
helper functions. So there are a few attr_accessors for the variables
like top, row_num, etc.
I think rather than have one Calendar class here, we can start to break
into Week and Day classes too and it should look a bit cleaner then.

Anything that was wrapping some content is now a method that takes
a block so that it is easier to mentally match up closing tags.